# docs/swagger.yaml
basePath: /api/v1
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  models.AuthUser:
    properties:
      password:
        example: '*****'
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  models.Author:
    properties:
      about:
        example: it's info about myself
        type: string
      avatar:
        example: filename.jpeg
        type: string
      email:
        example: admin@mail.ru
        type: string
      is_author:
        example: true
        type: boolean
      username:
        example: admin
        type: string
    required:
    - email
    - is_author
    - username
    type: object
  models.EmptyStruct:
    type: object
...
host: vdonate.ml:8080
info:
  contact:
    email: zeronethunter2001@gmail.com
    name: VDonate Support
  termsOfService: http://swagger.io/terms/
  title: VDonate API
  version: "1.0"
paths:
  /auth:
    get:
      description: Check authentication of `User` by cookies
      operationId: auth
      produces:
      - application/json
      responses:
        "200":
          description: Session was successfully found
          schema:
            $ref: '#/definitions/models.EmptyStruct'
        "401":
          description: User is unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User was not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: User authentication request
      tags:
      - auth
  /login:
    post:
      consumes:
      - application/json
      description: Authorization of `User`
      operationId: login
      parameters:
      - description: username and password
        in: body
        name: authData
        required: true
        schema:
          $ref: '#/definitions/models.AuthUser'
      produces:
      - application/json
      responses:
        "200":
          description: Session was successfully found
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Wrong login or password or bad data was received
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User was not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: User login request
      tags:
      - auth
  /logout:
    delete:
      description: Get request for user logout
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logout
          schema:
            $ref: '#/definitions/models.EmptyStruct'
        "400":
          description: Bad session / request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: User logout
      tags:
      - auth
...
  /users:
    post:
      consumes:
      - multipart/form-data
      description: Request to server for `User` creation
      operationId: signup
      parameters:
      - example: admin@mail.ru
        in: formData
        name: email
        required: true
        type: string
      - example: '*****'
        in: formData
        name: password
        required: true
        type: string
      - example: admin
        in: formData
        name: username
        required: true
        type: string
      - description: Upload avatar
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: User was successfully created
          schema:
            $ref: '#/definitions/models.EmptyStruct'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Username or email is already exists
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Creates a User
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    description: Authorization via Cookie
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"